backend/requirements.txt
# NDARite Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Document Generation
jinja2==3.1.2
weasyprint==60.2
python-docx==1.1.0

# External Integrations
stripe==7.8.0
docusign-esign==3.24.0
sendgrid==6.11.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1

# File Storage
boto3==1.34.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
